backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
azure-search-documents
openai
aiohttp
requests
google-auth
python-dotenv
python-multipart
google-api-python-client 
aiosqlite